Relocate main streamwatch
Moving StreamWatchListUpdateView to the streamwatch namespace
